+++ src/ui/setup_get_ingredients_window.py
@@ -9,7 +9,7 @@ from src.logger_handler import LoggerHandler
 from src.display_controller import DP_CONTROLLER
 from src.database_commander import DB_COMMANDER
 from src.machine.controller import MACHINE
-from src.bottles import set_fill_level_bars
+from src.tabs.bottles import set_fill_level_bars
 from src.dialog_handler import UI_LANGUAGE
 
 LOG_HANDLER = LoggerHandler("additional_ingredient", "production_logs")
